+++ markhub/views.py
@@ -139,7 +139,7 @@ class HomeView(TemplateView):
         user = self.request.user
         if user.is_authenticated and (g := get_github_handler(user)):
             context['repos'] = [
-                (repo.name, repo.created_at) 
+                (repo.name, repo.created_at, repo.private) 
                 for repo in g.get_user().get_repos() 
                 if user.username == repo.owner.login
             ]