+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - show private repos if you have private_repos permission
 
 ### Changed
+- rename Share button to Publish for private repos
+- refactor file actions in file-actions component
 
 ### Deprecated